Moved open source functions to own folder
